simplify Version string
Is absolutely useless to include any VCS specific version strings
or UUIDs into a version stirng: it makes it hard to normal people
to read and if built from a different repo they do not match anyway.
And if the VCS changes, all the bloat for nothing. People should
have learned this lesson a long time ago. What matters only is the
release tag: [0-9]+.[0-9]+.[0-9]+(.[0-9]+)*