new feature: --compact
HELP and TYPE comments are usually useless for DBs as wel as utilities
processing timeseries. So there is a lot (~ 50% sometimes even more) of
overhead sent over the wire without any needs. Producing and processing
this bloat needs additional ressources. So allow to run it e.g. in
production environments in an efficient way and skip sending this info.